+++ Clockwork/DataSource/DoctrineDataSource.php
@@ -5,43 +5,26 @@ use Clockwork\Request\Request;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Logging\SQLLogger;
 use Doctrine\ORM\EntityManager;
-use Illuminate\Log\Writer;
 
 class DoctrineDataSource extends DataSource implements SQLLogger
 {
 	/**
 	 * Internal array where queries are stored
-	 * @var array
 	 */
 	protected $queries = [];
 
 	/**
-	 * Doctrine event manager
-	 *
-	 * @var EventManager
-	 */
-	protected $evm;
-
-	/**
 	 * Doctrine entity manager
-	 *
-	 * @var EntityManager
 	 */
 	protected $enm;
 
 	/**
-	 * @var Logger
-	 */
-	protected $logger;
-
-	/**
 	 * For timing queries:
-	 * @var float|null
 	 */
 	public $start = null;
 
 	/**
-	 * @var null
+	 * Current recorded query
 	 */
 	public $query = null;
 
@@ -52,11 +35,9 @@ class DoctrineDataSource extends DataSource implements SQLLogger
 	 */
 	protected $connection;
 
-	public function __construct(EntityManager $enm, Writer $logger)
+	public function __construct(EntityManager $enm)
 	{
-		$this->evm = $enm->getEventManager();     //Unused, for now
-		$this->enm = $enm;
-		$this->logger = $logger;
+		$this->enm        = $enm;
 		$this->connection = $enm->getConnection();
 
 		$enm->getConnection()->getConfiguration()->setSQLLogger($this);
@@ -72,8 +53,8 @@ class DoctrineDataSource extends DataSource implements SQLLogger
 	public function startQuery($sql, array $params = null, array $types = null)
 	{
 		$this->start = microtime(true);
-		$sql = $this->replaceParams($sql, $params);
-		$sql = $this->formatQuery($sql);
+		$sql         = $this->replaceParams($sql, $params);
+		$sql         = $this->formatQuery($sql);
 
 		$this->query = [ 'sql' => $sql, 'params' => $params, 'types' => $types ];
 	}
@@ -81,10 +62,9 @@ class DoctrineDataSource extends DataSource implements SQLLogger
 	protected function formatQuery($sql)
 	{
 		$keywords = [ 'select', 'insert', 'update', 'delete', 'where', 'from', 'limit', 'is', 'null', 'having', 'group by', 'order by', 'asc', 'desc' ];
-		$regexp = '/\b' . implode('\b|\b', $keywords) . '\b/i';
+		$regexp   = '/\b' . implode('\b|\b', $keywords) . '\b/i';
 
-		$sql = preg_replace_callback($regexp, function($match)
-		{
+		$sql = preg_replace_callback($regexp, function($match){
 			return strtoupper($match[0]);
 		}, $sql);
 
@@ -98,7 +78,7 @@ class DoctrineDataSource extends DataSource implements SQLLogger
 
 			foreach ($params as $param) {
 				$param = $this->convertParam($param);
-				$sql = preg_replace($pattern, "\"$param\"", $sql, 1);
+				$sql   = preg_replace($pattern, "\"$param\"", $sql, 1);
 			}
 		}
 
@@ -108,12 +88,12 @@ class DoctrineDataSource extends DataSource implements SQLLogger
 	protected function convertParam($param)
 	{
 		if (is_object($param)) {
-			if (!method_exists($param, '__toString')) {
+			if (! method_exists($param, '__toString')) {
 				// Carbon Object
-				if (is_a($param, 'DateTime')) {
+				if (is_a($param, 'DateTime') || is_a($param, 'DateTimeImmutable')) {
 					$param = $param->format('Y-m-d');
 				} else {
-					throw new \Exception('Unstringable Object');
+					throw new \Exception('Unstringable Object: ' . get_class($param));
 				}
 			}
 		}
@@ -128,17 +108,7 @@ class DoctrineDataSource extends DataSource implements SQLLogger
 	{
 		$endTime = microtime(true) - $this->start;
 
-		/**
-		 * TODO Improve Connection name
-		 * For now, $this->connection->getDatabase() return the name of the Database
-		 */
-
 		$this->registerQuery($this->query['sql'], $this->query['params'], $endTime, $this->connection->getDatabase());
-
-		/**
-		 * This call is helpful to register queries in log file when the application is in debug/local mode
-		 */
-		$this->logger->debug($this->query['sql']. " [Executed in " . $endTime . " secs.] ", [ 'params' => $this->query['params'], 'types' => $this->query['types'] ]);
 	}
 
 	/**
@@ -171,6 +141,7 @@ class DoctrineDataSource extends DataSource implements SQLLogger
 	protected function getDatabaseQueries()
 	{
 		$queries = [];
+
 		foreach ($this->queries as $query) {
 			$queries[] = [
 				'query'      => $query['query'],