+++ Clockwork/Support/Lumen/ClockworkServiceProvider.php
@@ -1,59 +1,51 @@
-<?php namespace Clockwork\Support\Laravel;
+<?php namespace Clockwork\Support\Lumen;
 
 use Clockwork\Clockwork;
-use Clockwork\DataSource\EloquentDataSource;
-use Clockwork\DataSource\LaravelDataSource;
 use Clockwork\DataSource\PhpDataSource;
+use Clockwork\DataSource\LumenDataSource;
+use Clockwork\DataSource\EloquentDataSource;
 use Clockwork\DataSource\SwiftDataSource;
 
-use Illuminate\Foundation\Application;
+use Illuminate\Support\Facades\Facade;
 use Illuminate\Support\ServiceProvider;
-use Illuminate\Support\Str;
 
 class ClockworkServiceProvider extends ServiceProvider
 {
 	public function boot()
 	{
+		if ($this->isRunningWithFacades() && !class_exists('Clockwork')) {
+			class_alias('Clockwork\Support\Lumen\Facade', 'Clockwork');
+		}
+
 		if (!$this->app['clockwork.support']->isCollectingData()) {
 			return; // Don't bother registering event listeners as we are not collecting data
 		}
 
-		$this->app['clockwork.eloquent']->listenToEvents();
+		if ($this->app['clockwork.support']->isCollectingDatabaseQueries()) {
+			$this->app['clockwork.eloquent']->listenToEvents();
+		}
 
-		// create the clockwork instance so all data sources are initialized at this point
-		$this->app->make('clockwork');
+		if ($this->app['clockwork.support']->isCollectingEmails()) {
+			$this->app->make('clockwork.swift');
+		}
 
 		if (!$this->app['clockwork.support']->isEnabled()) {
 			return; // Clockwork is disabled, don't register the route
 		}
 
-		if ($this->isLegacyLaravel()) {
-			$this->app['router']->get('/__clockwork/{id}', 'Clockwork\Support\Laravel\Controllers\LegacyController@getData')->where('id', '[0-9\.]+');
-		} elseif ($this->isOldLaravel()) {
-			$this->app['router']->get('/__clockwork/{id}', 'Clockwork\Support\Laravel\Controllers\OldController@getData')->where('id', '[0-9\.]+');
-		} else {
-			$this->app['router']->get('/__clockwork/{id}', 'Clockwork\Support\Laravel\Controllers\CurrentController@getData')->where('id', '[0-9\.]+');
-		}
+		$this->app->get('/__clockwork/{id}', 'Clockwork\Support\Lumen\Controller@getData');
 	}
 
 	public function register()
 	{
-		if ($this->isLegacyLaravel() || $this->isOldLaravel()) {
-			$this->package('itsgoingd/clockwork', 'clockwork', __DIR__);
-		} else {
-			$this->publishes([ __DIR__ . '/config/clockwork.php' => config_path('clockwork.php') ]);
-		}
-
 		$this->app->singleton('clockwork.support', function($app)
 		{
-			$legacy = $this->isLegacyLaravel() || $this->isOldLaravel();
-
-			return new ClockworkSupport($app, $legacy);
+			return new ClockworkSupport($app);
 		});
 
-		$this->app->singleton('clockwork.laravel', function($app)
+		$this->app->singleton('clockwork.lumen', function($app)
 		{
-			return new LaravelDataSource($app);
+			return new LumenDataSource($app);
 		});
 
 		$this->app->singleton('clockwork.swift', function($app)
@@ -76,11 +68,14 @@ class ClockworkServiceProvider extends ServiceProvider
 
 			$clockwork
 				->addDataSource(new PhpDataSource())
-				->addDataSource($app['clockwork.laravel'])
-				->addDataSource($app['clockwork.swift']);
+				->addDataSource($app['clockwork.lumen']);
 
 			if ($app['clockwork.support']->isCollectingDatabaseQueries()) {
 				$clockwork->addDataSource($app['clockwork.eloquent']);
+			}
+
+			if ($app['clockwork.support']->isCollectingEmails()) {
+				$clockwork->addDataSource($app['clockwork.swift']);
 			}
 
 			foreach ($app['clockwork.support']->getAdditionalDataSources() as $name => $callable) {
@@ -92,11 +87,11 @@ class ClockworkServiceProvider extends ServiceProvider
 			return $clockwork;
 		});
 
-		$this->app['clockwork.laravel']->listenToEvents();
+		$this->app['clockwork.lumen']->listenToEvents();
 
 		// set up aliases for all Clockwork parts so they can be resolved by the IoC container
-		$this->app->alias('clockwork.support', 'Clockwork\Support\Laravel\ClockworkSupport');
-		$this->app->alias('clockwork.laravel', 'Clockwork\DataSource\LaravelDataSource');
+		$this->app->alias('clockwork.support', 'Clockwork\Support\Lumen\ClockworkSupport');
+		$this->app->alias('clockwork.lumen', 'Clockwork\DataSource\LumenDataSource');
 		$this->app->alias('clockwork.swift', 'Clockwork\DataSource\SwiftDataSource');
 		$this->app->alias('clockwork.eloquent', 'Clockwork\DataSource\EloquentDataSource');
 		$this->app->alias('clockwork', 'Clockwork\Clockwork');
@@ -106,15 +101,6 @@ class ClockworkServiceProvider extends ServiceProvider
 		if ($this->app['clockwork.support']->getConfig('register_helpers', true)) {
 			require __DIR__ . '/helpers.php';
 		}
-
-		if ($this->isLegacyLaravel()) {
-			$this->app->middleware('Clockwork\Support\Laravel\ClockworkLegacyMiddleware', [ $this->app ]);
-		} elseif ($this->isOldLaravel()) {
-			$this->app['router']->after(function($request, $response)
-			{
-				return $this->app['clockwork.support']->process($request, $response);
-			});
-		}
 	}
 
 	/**
@@ -123,7 +109,9 @@ class ClockworkServiceProvider extends ServiceProvider
 	public function registerCommands()
 	{
 		// Clean command
-		$this->app->bind('command.clockwork.clean', 'Clockwork\Support\Laravel\ClockworkCleanCommand');
+		$this->app['command.clockwork.clean'] = $this->app->share(function($app){
+			return new ClockworkCleanCommand();
+		});
 
 		$this->commands(
 			'command.clockwork.clean'
@@ -135,13 +123,8 @@ class ClockworkServiceProvider extends ServiceProvider
 		return [ 'clockwork' ];
 	}
 
-	public function isLegacyLaravel()
+	protected function isRunningWithFacades()
 	{
-		return Str::startsWith(Application::VERSION, [ '4.1', '4.2' ]);
-	}
-
-	public function isOldLaravel()
-	{
-		return Str::startsWith(Application::VERSION, '4.0');
+		return Facade::getFacadeApplication() !== null;
 	}
 }